lazos-api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
pgvector==0.2.4

# Pydantic and validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Image processing
Pillow==10.2.0

# Storage (Cloudflare R2 / S3-compatible)
boto3==1.34.18

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
